KNOX-2226 - Addressed review comments - round 4
