+++ gateway-provider-security-jwt/src/main/java/org/apache/knox/gateway/provider/federation/jwt/filter/SSOCookieFederationFilter.java
@@ -17,7 +17,6 @@
  */
 package org.apache.knox.gateway.provider.federation.jwt.filter;
 
-import org.apache.commons.lang3.StringUtils;
 import org.apache.knox.gateway.config.GatewayConfig;
 import org.apache.knox.gateway.i18n.messages.MessagesFactory;
 import org.apache.knox.gateway.provider.federation.jwt.JWTMessages;
@@ -94,16 +93,20 @@ public class SSOCookieFederationFilter extends AbstractJWTFilter {
     }
 
     // gateway path for deriving an idp url when missing
+    setGatewayPath(filterConfig);
+
+    configureExpectedParameters(filterConfig);
+  }
+
+  private void setGatewayPath(FilterConfig filterConfig) {
     gatewayPath = filterConfig.getInitParameter(GATEWAY_PATH);
-    if (StringUtils.isBlank(gatewayPath)) {
+    if (gatewayPath == null || gatewayPath.isEmpty()) {
       final GatewayConfig gatewayConfig = filterConfig.getServletContext() == null ? null
           : (GatewayConfig) filterConfig.getServletContext().getAttribute(GatewayConfig.GATEWAY_CONFIG_ATTRIBUTE);
       if (gatewayConfig != null) {
         gatewayPath = gatewayConfig.getGatewayPath();
       }
     }
-
-    configureExpectedParameters(filterConfig);
   }
 
   @Override
