+++ gateway-release/src/assembly.xml
@@ -114,7 +114,6 @@
             <outputDirectory>lib</outputDirectory>
             <includes>
                 <include>org.apache.knox:gateway-*</include>
-                <include>org.apache.knox:knox-homepage-backend</include>
                 <!-- include>org.apache.knox:gateway-release-*</include -->
             </includes>
             <excludes>
