+++ gateway-service-metadata/src/main/java/org/apache/knox/gateway/service/metadata/GeneralProxyInformationMarshaller.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.knox.homepage;
+package org.apache.knox.gateway.service.metadata;
 
 import java.io.IOException;
 import java.io.OutputStream;
