+++ gateway-service-metadata/src/main/resources/META-INF/services/org.apache.knox.gateway.deploy.ServiceDeploymentContributor
@@ -15,4 +15,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 ##########################################################################
-org.apache.knox.homepage.KnoxMetadataServiceDeploymentContributor
+org.apache.knox.gateway.service.metadata.KnoxMetadataServiceDeploymentContributor
